requirements.txt
Django==4.2.*
djangorestframework==3.14.*
djoser==2.2.*
psycopg2-binary==2.9.*
gunicorn==21.2.*
django-cors-headers==4.3.*
python-dotenv==1.0.*
whitenoise==6.6.* # Added for static files in production
